Clean up some example warnings

+++ libraries/WebServer/examples/AdvancedWebServer/AdvancedWebServer.ino
@@ -97,7 +97,7 @@ void handleNotFound() {
   message += server.args();
   message += "\n";
 
-  for (uint8_t i = 0; i < server.args(); i++) {
+  for (int i = 0; i < server.args(); i++) {
     message += " " + server.argName(i) + ": " + server.arg(i) + "\n";
   }
